requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.3; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
djangorestframework==3.15.2; python_version >= '3.8'
mysql-connector-python==9.1.0; python_version >= '3.9'
pillow==11.0.0; python_version >= '3.9'
python-dotenv==1.0.1; python_version >= '3.8'
sqlparse==0.5.2; python_version >= '3.8'
